Fix for calculated dae_wmk in DRI+FIFO[continued_2]

+++ akm_src/slpi_proc/ssc/sensors/ak0991x/src/sns_ak0991x_sensor_instance.h
@@ -153,7 +153,7 @@ typedef struct ak0991x_config_event_info
 {
   ak0991x_mag_odr odr;
   uint16_t         fifo_wmk;
-  uint16_t         dae_wmk;
+  uint32_t         dae_wmk;
 } ak0991x_config_event_info;
 
 typedef struct ak0991x_mag_info
